Introduce a way to mix generated datasets before sending to training
Rebased from
https://github.com/aakankshaduggal/sdg/pull/4
https://github.com/aakankshaduggal/sdg/pull/13
https://github.com/aakankshaduggal/sdg/pull/14

Refactored and changed by me to be as least disruptive as possible in
merging in, which means leaving the legacy train and messages outputs in
place in generated_data.py.

This introduces a new Recipe class that references multiple generated
datasets and has logic to combine those into a single mixed dataset
based on a configurable number / ratio of samples to take from each
dataset.

This adds new output artifacts for each data generation run:
* a `node_datasets_*` subfolder that contains the raw generated samples
  from each taxonomy leaf node as a set of jsonl files - within here,
  each skill taxonomy node gets a single jsonl and knowledge nodes get
  two jsonl files used in separate downstream steps of training
* a `skills_train_msgs_*.jsonl` and a `knowledge_train_msgs_*.jsonl`
  file that contains mixtures of the above raw generated samples based
  on proportions specified during the mixing process

Co-authored-by: shivchander <shivchander.s30@gmail.com>
Co-authored-by: Khaled Sulayman <khaled@thesulaymans.com>
Co-authored-by: abhi1092 <abhi1092@gmail.com>
Co-authored-by: Aakanksha Duggal <aduggal@redhat.com>
Signed-off-by: Ben Browning <bbrownin@redhat.com>